Refactor testing to allow running with fprime-util
