+++ FppTestProject/CMakeLists.txt
@@ -10,7 +10,11 @@ project(FppTest C CXX)
 
 include(CheckCXXCompilerFlag)
 set(BUILD_TESTING ON)
-set(FPRIME_ENABLE_UT_COVERAGE ON)
+
+if (FPRIME_ENABLE_DIRECT_PORT_CALLS)
+    set(FPRIME_ENABLE_FRAMEWORK_UTS OFF)
+    add_compile_options(-DFW_DIRECT_PORT_CALLS=1)
+endif ()
 
 include("${CMAKE_CURRENT_LIST_DIR}/../cmake/FPrime.cmake")
