+++ FppTestProject/FppTest/CMakeLists.txt
@@ -3,6 +3,13 @@
 #
 # Builds unit tests for FPP autocoder
 ###
+
+if (FPRIME_ENABLE_DIRECT_PORT_CALLS)
+    # Disable unit testing for components that require
+    # indirect port calls
+    set(__FPRIME_NO_UT_GEN__ ON)
+endif ()
+
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/array/")
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/component/")
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/dp/")
@@ -10,6 +17,12 @@ add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/state_machine/")
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/enum/")
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/interfaces/")
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/struct/")
+
+if (FPRIME_ENABLE_DIRECT_PORT_CALLS)
+    # Re-enable unit testing for the topology since it uses direct port calls
+    set(__FPRIME_NO_UT_GEN__ OFF)
+endif ()
+
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/topology/")
 
 set(SOURCE_FILES "source.cpp")
