+++ FppTestProject/FppTest/topology/main/CMakeLists.txt
@@ -14,3 +14,15 @@ set(MOD_DEPS
 )
 
 register_fprime_module()
+
+# Declare dependencies on test modules
+set(UT_MOD_DEPS
+        STest
+)
+
+set(UT_SOURCE_FILES
+        "${CMAKE_CURRENT_LIST_DIR}/main.cpp"
+        "${CMAKE_CURRENT_LIST_DIR}/../../component/types/FormalParamTypes.cpp"
+        "${CMAKE_CURRENT_LIST_DIR}/../../utils/Utils.cpp"
+)
+register_fprime_ut()
